pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["phbui <phuc.bui@agoda.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
boto3 = "^1.35.59"
python-multipart = "^0.0.17"
python-dotenv = "^1.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
cffi = "^1.17.1"
paho-mqtt = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
